Removed ui directory. Moved PagerIndicatorValues.kt to about directory. My line of thinking is that pagerindicator directory is its own feature and doesn't rely on PagerIndicatorValues.kt.
